Rename folder to camelCase
